Rename evaluateFhirDataService to encounterInfoService. Change Encounter component name to Encounter Info to match Viewer. Update imports and props.
