+++ containers/ecr-viewer/src/app/view-data/components/EcrDocument/accordion-items.tsx
@@ -7,7 +7,7 @@ import { toKebabCase } from "@/app/utils/format-utils";
 import ClinicalInfo from "@/app/view-data/components/ClinicalInfo";
 import Demographics from "@/app/view-data/components/Demographics";
 import EcrMetadata from "@/app/view-data/components/EcrMetadata";
-import EncounterDetails from "@/app/view-data/components/Encounter";
+import EncounterInfo from "@/app/view-data/components/EncounterInfo";
 import LabInfo from "@/app/view-data/components/LabInfo";
 import PregnancyInfo from "@/app/view-data/components/PregnancyInfo";
 import SocialHistory from "@/app/view-data/components/SocialHistory";
@@ -18,7 +18,7 @@ import {
   evaluateProviderData,
   evaluateFacilityData,
   evaluateHospitalEncounterData,
-} from "@/app/view-data/services/evaluateFhirDataService";
+} from "@/app/view-data/services/encounterInfoService";
 import { evaluateSocialData } from "@/app/view-data/services/socialHistoryService";
 import { evaluateDemographicsData } from "@/app/view-data/services/demographicsService";
 import { evaluatePregnancyData } from "@/app/view-data/services/pregnancyInfoService";
@@ -168,7 +168,7 @@ export const getEcrDocumentAccordionItems = (
           hospitalEncounterData.availableData.length > 0 ||
           facilityData.availableData.length > 0 ||
           providerData.availableData.length > 0 ? (
-            <EncounterDetails
+            <EncounterInfo
               encounterData={encounterData.availableData}
               hospitalEncounterData={hospitalEncounterData.availableData}
               facilityData={facilityData.availableData}
