+++ containers/ecr-viewer/src/app/view-data/components/EncounterInfo.tsx
@@ -7,7 +7,7 @@ import {
 
 import { DataDisplay, DataTableDisplay, DisplayDataProps } from "./DataDisplay";
 
-interface EncounterProps {
+interface EncounterInfoProps {
   encounterData: DisplayDataProps[];
   hospitalEncounterData: DisplayDataProps[];
   facilityData: DisplayDataProps[];
@@ -15,20 +15,20 @@ interface EncounterProps {
 }
 
 /**
- * Functional component for displaying encounter details.
- * @param props - Props containing encounter details.
+ * Functional component for displaying Encounter Info section.
+ * @param props - Props containing all encounter details.
  * @param props.encounterData - Encounter details to be displayed.
  * @param props.hospitalEncounterData - Hospital Encounter Diagnosis details to be displayed.
  * @param props.providerData - Provider details to be displayed.
  * @param props.facilityData - Facility details to be displayed.
  * @returns The JSX element representing the encounter details.
  */
-const EncounterDetails = ({
+const EncounterInfo = ({
   encounterData,
   hospitalEncounterData,
   facilityData,
   providerData,
-}: EncounterProps) => {
+}: EncounterInfoProps) => {
   return (
     <AccordionSection>
       <EncounterSection title="Encounter Details" data={encounterData} />
@@ -73,4 +73,4 @@ const EncounterSection = ({
     )
   );
 };
-export default EncounterDetails;
+export default EncounterInfo;
